Add associations. Neat!
Register associations with the following methods ($model is the class
of the association, $class is the class you're associating from):
* Model::one_to_many($model, $name, $class);
* Model::many_to_many($model, $name, $class);
* Model::many_to_one($model, $name, $class);

Access associations from an instance of a Model like follows:
* $association = $model->load_association($name);
* $association = $model->{"load_$name"}(); // also valid

The loaded associations will either be an array (for one_to_many or
many_to_many), or a single instance of the associated model class (for
many_to_one). The associations are not currently reciprocal, but they
are caching. To force a reload of an association, call it like this:
* $association = $model->load_association($name, TRUE);
* $association = $model->{"load_$name"}(TRUE);

This update also adds Model::sort(), which can be used to sort arrays
of Model objects by ID. Would be called like so:
* usort($array, array('Model', 'sort'));

Finally, this fixes a (sort of) bug which made the set_* methods
somewhat case-insensitive. That is, Set_*, sEt_*, etc. were all
previously valid. Now only the all-lowercase form is valid.