Streamline file subtest labeling.
Rather than printing the filename and test count in the parsing
function, the test counts are printed as part of the test name. This
cuts down on lines and excess noise significantly. It turns out, it was
unnecessary to print pass/all once the is() function printed notations
to that effect automatically.