Strip out database reliance for quoting.
It was proving difficult to test methods when the underlying method
would not work without a live database connection. While it's nice to
use the PostgreSQL function as it's coded by people smarter than I am,
it's also less practical. I remain hopeful that the logic is simple
enough that I can successfully emulate it safely.